simplify & rationalize some things
