+++ lib/FragmentLoaderClassProvider.js
@@ -7,7 +7,7 @@ import HTTPRequest from '../dashjs/src/streaming/vo/metrics/HTTPRequest';
 import SegmentView from './SegmentView';
 import TrackView from './TrackView';
 
-function FragmentLoaderClassProvider(downloaderProvider) {
+function FragmentLoaderClassProvider(wrapper) {
 
     const FRAGMENT_LOADER_ERROR_LOADING_FAILURE = 1;
     const FRAGMENT_LOADER_ERROR_NULL_REQUEST = 2;
@@ -26,13 +26,13 @@ function FragmentLoaderClassProvider(downloaderProvider) {
         const metricsModel = config.metricsModel;
 
         let instance,
-            srLoader,
+            peerAgent,
             _abort,
             retryTimers,
             downloadErrorToRequestTypeMap;
 
         function setup() {
-            srLoader = downloaderProvider.srDownloader;
+            peerAgent = wrapper.peerAgent;
             retryTimers = [];
 
             downloadErrorToRequestTypeMap = {
@@ -202,7 +202,7 @@ function FragmentLoaderClassProvider(downloaderProvider) {
                 }
             };
 
-            _abort = srLoader.getSegment(
+            _abort = peerAgent.getSegment(
                 srRequest,
                 {
                     onSuccess,
