+++ lib/ManifestHelper.js
@@ -1,15 +1,16 @@
 import TrackView from './TrackView';
 import SegmentsGetter from '../dashjs/src/dash/utils/SegmentsGetter';
 import SegmentsCache from './SegmentsCache';
+import DashjsInternals from './DashjsInternals';
 
 class ManifestHelper {
 
-    constructor (player, manifestProvider, dashjsInternals) {
+    constructor (manifestProvider) {
         this._id = new Date().toString();
         // console.info("ManifestHelper(): id=", this._id);
-        this._player = player;
         this._manifestProvider = manifestProvider;
-        this._dashjsInternals = dashjsInternals;
+        let player = this._player = manifestProvider.player;
+        this._dashjsInternals = new DashjsInternals(player);
         this._segmentsCache = new SegmentsCache(player);
     }
