+++ scripts/dist_bundle
@@ -6,4 +6,4 @@ DIST_BUNDLE = "dist/bundle/dashjs-p2p-bundle.js"
 
 `mkdir -p dist`
 `mkdir -p dist/bundle`
-system("#{NODE_BIN}/browserify --debug -p browserify-derequire -t [babelify] -s DashjsP2PBundle lib/DashjsMediaPlayerBundle.js | #{NODE_BIN}/exorcist #{DIST_BUNDLE}.map -b . > #{DIST_BUNDLE}")
+system("#{NODE_BIN}/browserify --debug -p browserify-derequire -t [babelify] -s DashjsP2PBundle lib/DashjsBundle.js | #{NODE_BIN}/exorcist #{DIST_BUNDLE}.map -b . > #{DIST_BUNDLE}")
